Home Work

+++ evgeniy-vor/src/main/hw/hw01/Solution.java
@@ -105,36 +105,36 @@ public class Solution {
         //TaskBonus
         Scanner in = new Scanner(System.in);
         System.out.print("Please enter the number: ");
-        int getNumberFromConsole  = in.nextInt();
+        byte getNumberFromConsole  = in.nextByte();
 
         BonusHomeWork.convertNumber(getNumberFromConsole);
     }
 
     static class BonusHomeWork {
 
-        public static void convertNumber(int b){
-            int a1 = b%2;
-            int b1 = b/2;
+        public static void convertNumber(byte b){
+            byte a1 = (byte)(b%2);
+            byte b1 = (byte)(b/2);
 
-            int a2 = b1%2;
-            int b2 = b1/2;
+            byte a2 = (byte)(b1%2);
+            byte b2 = (byte)(b1/2);
 
-            int a3 = b2%2;
-            int b3 = b2/2;
+            byte a3 = (byte)(b2%2);
+            byte b3 = (byte)(b2/2);
 
-            int a4 = b3%2;
-            int b4 = b3/2;
+            byte a4 = (byte)(b3%2);
+            byte b4 = (byte)(b3/2);
 
-            int a5 = b4%2;
-            int b5 = b4/2;
+            byte a5 = (byte)(b4%2);
+            byte b5 = (byte)(b4/2);
 
-            int a6 = b5%2;
-            int b6 = b5/2;
+            byte a6 = (byte)(b5%2);
+            byte b6 = (byte)(b5/2);
 
-            int a7 = b6%2;
-            int b7 = b6/2;
+            byte a7 = (byte)(b6%2);
+            byte b7 = (byte)(b6/2);
 
-            int a8 = b7%2;
+            byte a8 = (byte)(b7%2);
 
             //Вывод в консоль:
             System.out.println(a8+" "+a7+" "+a6+" "+a5+" "+a4+" "+a3+" "+a2+" "+a1);
